A simpler pass manager (#3286)
* split!

* passes()

* max_iteration

* replace

* some RunningPassManager clean up

* callback support

* property_set result

* lint

* cleaning up runninpassmanager

* docstring

* args

* runningpassmanager style

* fix the drawer with the new format from passes

* new refs

* Update qiskit/visualization/pass_manager_visualization.py


* no label for linear

* refs

* _pass_sets

* pass set in a docstring

* lint
