+++ qiskit/visualization/pass_manager_visualization.py
@@ -105,11 +105,12 @@ def pass_manager_drawer(pass_manager, filename, style=None, raw=False):
 
     for index, controller_group in enumerate(passes):
 
-        # label is the name of the flow controller (without the word controller)
-        label = "[%s] %s" % (index, controller_group['type'].__name__.replace('Controller', ''))
+        # label is the name of the flow controller parameter
+        label = "[%s] %s" % (index, ', '.join(controller_group['flow_controllers']))
 
         # create the subgraph for this controller
-        subgraph = pydot.Cluster(str(component_id), label=label, fontname='helvetica')
+        subgraph = pydot.Cluster(str(component_id), label=label, fontname='helvetica',
+                                 labeljust='l')
         component_id += 1
 
         for pass_ in controller_group['passes']:
